+++ minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/item/DynamicAttributeItem.java
@@ -56,6 +56,8 @@ public interface DynamicAttributeItem extends DynamicAttributeTool, DynamicDurab
     UUID TEST_UUID = UUID.fromString("CB3F55D3-645C-4F38-A497-9C13A34DB5CF");
     UUID ADDITION_ATTACK_DAMAGE_MODIFIER_ID = UUID.fromString("CB3F55D5-655C-4F38-A497-9C13A33DB5CF");
 
+    UUID ADDITION_HEALTH_MODIFIER_ID = UUID.randomUUID();
+
     UUID ADDITION_LUCK_MODIFIER_ID = UUID.fromString("CC3F55D5-755C-4F38-A497-9C13A33DB5CF");
 
     UUID ADDITION_ARMOR_MODIFIER_ID = UUID.fromString("AC3F55D5-755C-4F38-A497-9C13A63DB5CF");
@@ -123,12 +125,18 @@ public interface DynamicAttributeItem extends DynamicAttributeTool, DynamicDurab
         //Attack damage addition
         builder.put(EntityAttributes.GENERIC_ATTACK_DAMAGE, new EntityAttributeModifier(ADDITION_ATTACK_DAMAGE_MODIFIER_ID, "Attack Damage Addition", currentToolDamage - baseToolDamage, EntityAttributeModifier.Operation.ADDITION));
 
+        //Additional armor
         int luck = LuckHandler.of(dynamicProperties(stack)).map(Attribute::asInt).orElse(0);
         builder.put(EntityAttributes.GENERIC_LUCK, new EntityAttributeModifier(ADDITION_LUCK_MODIFIER_ID, "Luck addition", luck, EntityAttributeModifier.Operation.ADDITION));
 
+        //Additional armor
         float armor = Armor.of(dynamicProperties(stack)).asFloat();
         builder.put(EntityAttributes.GENERIC_ARMOR, new EntityAttributeModifier(ADDITION_ARMOR_MODIFIER_ID, "Armor addition", armor, EntityAttributeModifier.Operation.ADDITION));
-        
+
+        //Additional health
+        int additionalHealth = AdditionalHealthHandler.of(dynamicProperties(stack)).map(Attribute::asInt).orElse(0);
+        builder.put(EntityAttributes.GENERIC_MAX_HEALTH, new EntityAttributeModifier(ADDITION_HEALTH_MODIFIER_ID, "Health addition", additionalHealth, EntityAttributeModifier.Operation.ADDITION));
+
         //Attack speed
         float baseAttackSpeed = AttackSpeed.apply(dynamicProperties(stack), target);
         float currentAttackSpeed = AttackSpeed.apply(defaultProperties());