+++ changelog.md
@@ -8,6 +8,7 @@ you have backed up your save.**
 ## Changelog 0.10.7
 
 * Added armor property
+* Added The max health addition property
 * Added luck property
 * Added critical chance property
 * Fixed simple sword guard recipe to not conflict with wooden bowl